+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/map/MapIOUtils.java
@@ -1,11 +1,6 @@
 package dev.ftb.mods.ftbchunks.client.map;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
+import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.zip.DeflaterOutputStream;