+++ common/src/main/java/dev/ftb/mods/ftbchunks/data/FTBChunksTeamData.java
@@ -9,6 +9,7 @@ import dev.ftb.mods.ftbchunks.net.ChunkSendingUtils;
 import dev.ftb.mods.ftbchunks.net.SendChunkPacket;
 import dev.ftb.mods.ftbchunks.net.SendGeneralDataPacket;
 import dev.ftb.mods.ftbchunks.net.SendManyChunksPacket;
+import dev.ftb.mods.ftbchunks.util.DimensionFilter;
 import dev.ftb.mods.ftblibrary.math.ChunkDimPos;
 import dev.ftb.mods.ftblibrary.snbt.SNBT;
 import dev.ftb.mods.ftblibrary.snbt.SNBTCompoundTag;
@@ -54,6 +55,7 @@ public class FTBChunksTeamData {
 	public static final PrivacyProperty ENTITY_INTERACT_MODE = new PrivacyProperty(new ResourceLocation(FTBChunks.MOD_ID, "entity_interact_mode"), PrivacyMode.ALLIES);
 	public static final PrivacyProperty NONLIVING_ENTITY_ATTACK_MODE = new PrivacyProperty(new ResourceLocation(FTBChunks.MOD_ID, "nonliving_entity_attack_mode"), PrivacyMode.ALLIES);
 	public static final BooleanProperty ALLOW_EXPLOSIONS = new BooleanProperty(new ResourceLocation(FTBChunks.MOD_ID, "allow_explosions"), false);
+	public static final BooleanProperty ALLOW_MOB_GRIEFING = new BooleanProperty(new ResourceLocation(FTBChunks.MOD_ID, "allow_mob_griefing"), false);
 	public static final PrivacyProperty CLAIM_VISIBILITY = new PrivacyProperty(new ResourceLocation(FTBChunks.MOD_ID, "claim_visibility"), PrivacyMode.PUBLIC);
 
 	//	public static final PrivacyProperty MINIMAP_MODE = new PrivacyProperty(new ResourceLocation(FTBChunks.MOD_ID, "minimap_mode"), PrivacyMode.ALLIES);
@@ -149,7 +151,7 @@ public class FTBChunksTeamData {
 
 		if (chunk != null) {
 			return ClaimResults.ALREADY_CLAIMED;
-		} else if (FTBChunksWorldConfig.CLAIM_DIMENSION_BLACKLIST_SET.contains(pos.dimension)) {
+		} else if (!DimensionFilter.isDimensionOK(pos.dimension)) {
 			return ClaimResults.DIMENSION_FORBIDDEN;
 		} else if (!team.getType().isServer() && getClaimedChunks().size() >= getMaxClaimChunks()) {
 			return ClaimResults.NOT_ENOUGH_POWER;
@@ -470,6 +472,10 @@ public class FTBChunksTeamData {
 
 	public boolean allowExplosions() {
 		return team.getProperty(ALLOW_EXPLOSIONS);
+	}
+
+	public boolean allowMobGriefing() {
+		return team.getProperty(ALLOW_MOB_GRIEFING);
 	}
 
 	public void setLastLoginTime(long when) {